feat(ui): refactor Settings into subpages layout
- Add Settings layout with sidebar navigation for subpages
- Create LLM Providers/Models page at /settings/providers
- Create API Keys page at /settings/api-keys
- Create Members page placeholder at /settings/members (needs APIs)
- Update main Settings page to redirect to providers
- Update sidebar API Keys link to use new route
