+++ apps/ui/src/components/layout/sidebar.tsx
@@ -113,7 +113,7 @@ export function Sidebar() {
                     <User className="mr-2 h-4 w-4" />
                     Profile
                   </DropdownMenuItem>
-                  <DropdownMenuItem onClick={() => router.push("/settings#api-keys")}>
+                  <DropdownMenuItem onClick={() => router.push("/settings/api-keys")}>
                     <Key className="mr-2 h-4 w-4" />
                     API Keys
                   </DropdownMenuItem>
